backend/pyproject.toml
[tool.poetry]
name = "budgeteer-api"
version = "0.1.0"
description = "Budgeteer backend (FastAPI)"
authors = ["Budgeteer"]
readme = "README.md"
packages = [{include="app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
python-multipart = "^0.0.9"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
SQLAlchemy = "^2.0.34"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
PyJWT = "^2.9.0"
email-validator = "^2.2.0"
python-dateutil = "^2.9.0.post0"
itsdangerous = "^2.2.0"
qrcode = "^7.4.2"
aiofiles = "^24.1.0"
slowapi = "^0.1.9"
python-crontab = "^3.0.0"
pendulum = "^3.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
black = "^24.8.0"
ruff = "^0.6.9"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]
